Create MongoDB client and better config.
